AI/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
chromadb
pydantic
requests
